test(img): migrate basic
